% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.R
\name{check_file_existance}
\alias{check_file_existance}
\title{For checking file existance}
\usage{
check_file_existance(input.file = NULL, error = TRUE,
  verbose = FALSE)
}
\arguments{
\item{input.file}{input sequence file}

\item{error}{If \code{TRUE}, throw an error if programs is not installed.}

\item{verbose}{If \code{TRUE}, print progress reports.}
}
\value{
provide logical


 @examples
 check_file_existance(input.file = NULL, error = TRUE, verbose = FALSE)
}
\description{
This function check the file paths.
}
