% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.R
\name{check_signalp_version}
\alias{check_signalp_version}
\title{For checking signalp version}
\usage{
check_signalp_version(path = NULL)
}
\arguments{
\item{path}{for path of the programe.}
}
\value{
provide version of the signalp programe


 @examples

check_signalp_version(path = NULL)
}
\description{
This function check check signalp version
.
}
