% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.R
\name{check_signalp_version_match}
\alias{check_signalp_version_match}
\title{For check signalp program and version match}
\usage{
check_signalp_version_match(program_version = NULL,
  input_version = NULL)
}
\arguments{
\item{program_version}{for version of the programe.}

\item{input_version}{Provided input version by user}
}
\value{
return logical condition


 @examples

check_signalp_version_match(program_version=NULL, input_version= NULL)
}
\description{
This function check signalp program and version match
.
}
