% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.R
\name{external_programe_existance}
\alias{external_programe_existance}
\title{For checking programe existance and set default path}
\usage{
external_programe_existance(programe = NULL, path = NULL,
  error = TRUE, verbose = FALSE)
}
\arguments{
\item{programe}{parameter for programe name.}

\item{path}{for path of the programe.}

\item{error}{If \code{TRUE}, throw an error if programs is not installed.}

\item{verbose}{If \code{TRUE}, print progress reports.}
}
\value{
provide path of the programe


 @examples

external_programe_existance (programe = NULL, path = NULL, error = TRUE, verbose = FALSE)
}
\description{
This function check the program and file paths.
}
