% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.R
\name{get_mafft}
\alias{get_mafft}
\title{MAFFT Alignment}
\usage{
get_mafft(alignment.file = NULL, sequences = NULL, mafft.path = NULL,
  num.threads = NULL)
}
\arguments{
\item{alignment.file}{MAFFT alignment resulted file}

\item{sequences}{PATT_REG sequence reading and stored in sequences}

\item{mafft.path}{complete path of MAFFT}

\item{num.threads}{Number of threads to be used by MAFFT.}
}
\value{
Resulted output file name


 @examples
get_mafft(alignment.file = NULL, sequences = NULL, mafft.path = NULL, num.threads = NULL)
}
\description{
This function is perfoms multiple sequence alignment when alignment file is not provided by user
}
