% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getorf.R
\name{get_orf}
\alias{get_orf}
\title{Find and extract open reading frames (ORFs)  from nucleotide sequences}
\usage{
get_orf(getorf.path = NULL, input.file = NULL, output.file = NULL,
  error = TRUE, verbose = FALSE)
}
\arguments{
\item{getorf.path}{Path of getorf excutable file path,path.getorf excutable file will provide along with the package,if path is null then this function will take default path from syste path.}

\item{input.file}{inputfile should have DNA query sequence filename.}

\item{output.file}{name of output filename.}

\item{error}{If \code{TRUE}, throw an error if programs is not installed.}

\item{verbose}{If \code{TRUE}, print progress reports.}
}
\value{
return output filename

 @examples

get_orf( getorf.path=NULL, input.file="filename", output.file = NULL)
}
\description{
This function finds and outputs the sequences of open reading frames (ORFs) in one or more nucleotide sequences
}
