% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetp.R
\name{get_targetp}
\alias{get_targetp}
\title{Predicts sub cellular location of eukaryotic proteins}
\usage{
get_targetp(targetp.path = NULL, organismgroup = NULL, file = NULL)
}
\arguments{
\item{targetp.path}{Localpath of folder containing target binary excutable file or the excutable file itself, if path not specified, then targetp must be in program search path.}

\item{organismgroup}{Please provide oraganismgroup for targetp: -P|-N-use plant/non-plant networks mandatory, -c-include cleavage site predictions, -h-print this note and exit,  -v-print ve           rsion info and exit, -p-chloroplast prediction cutoff, default, -s-secretory pathway prediction cutoff, default, -t-mitochondrial prediction cutoff, default, -o-other location pred           iction cutoff, default.}

\item{file}{Please provide fasta input file}
}
\value{
Returns dataframe which contains summary of targetp result file

 @examples

get_targetp(targetp.path = NULL, organismgroup = NULL, file = input_file)
}
\description{
This function predicts the sub cellular location of eukaryotic proteins. The location assignment is based on the predicted presence of any of the N-terminal presequences:chloroplast transit peptide (cTP), mitochondrial targeting peptide (mTP) or secretory pathway signal peptide (SP).
}
