% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thmm.R
\name{get_tmhmm}
\alias{get_tmhmm}
\title{Prediction of transmembrane helices in proteins}
\usage{
get_tmhmm(tmhmm.path = NULL, file = NULL)
}
\arguments{
\item{tmhmm.path}{Localpath of folder containing tmhmm binary excutable file or the excutable file itself, if path not specified, then tmhmm must be in program search path}

\item{file}{Provide input fasta file name}
}
\value{
Returns dataframe which contain summary of tmhmm result file
}
\description{
This function Predicts the presence of transmembrane helices and extract and returns protiens with no transmembrane helices
}
\examples{
 \dontrun{

tmhmm_data <- get_tmhmm(tmhmm.path = NULL, file = NULL)
}

}
