% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{hmm.plot}
\alias{hmm.plot}
\title{Plots the relative frequencies of each position for hmmsearch table.}
\usage{
hmm.plot(hmm_data = NULL)
}
\arguments{
\item{hmm_data}{The HMM profile table resulting from hmm.search}
}
\description{
This function plots the results from hmm.search as a pointplot with amino acids in the x axis and the relative frequency of each amino acid in the y axis
}
\examples{
\dontrun{

hmm.plot(hmm_data = NULL)
}
}
\keyword{PATT_REG}
\keyword{plot}
