% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmsearch.R
\name{hmm.search}
\alias{hmm.search}
\title{Searching for motifs using HMM searches}
\usage{
hmm.search(original.seq, regex.seq, alignment.file = NULL,
  save.alignment = FALSE, mafft.path = NULL, num.threads = 2,
  hmm.tresh = 0, hmm.path = NULL, seed = 12345)
}
\arguments{
\item{original.seq}{The absolute path for the original six-frame translation FASTA file}

\item{regex.seq}{A list[1] of objects returns from pattern.search. The HMM profile will be constructed using these sequences}

\item{alignment.file}{The absolute path for an alignment file of the sequences to build the hmmer profile from.It's recommended that the alignment file cointains the same sequences than the `regex.seq` files. If the user provides the absolute path, *motiFUN* won't use MAFFT to align the sequences and will use the alignment for the HMMER searches. If no alignment file is provided, *motifFUN* will use MAFFT to align the sequences from `regex.seq` and run HMMER.}

\item{save.alignment}{Save the alignment in the returning object. The MAFFT alignment will be saved as the first element of the returned object}

\item{mafft.path}{Local path of folder containing the MAFFT binary executable file or the executable file itself. If not specified, then MAFFT must be in the program search path.}

\item{num.threads}{Number of threads to be used by MAFFT.}

\item{hmm.tresh}{Set the bit score cutoff for the per-sequence ranked hit list to a real number. (Default = 0)}

\item{hmm.path}{Local path of  folder containing the HMMER binaries.  If not specified, then HMMER executables must be in the program search path.}

\item{seed}{The seed to used with HMMER commands. Set this to get the same output each time.}
}
\value{
Returns total sequence.

 @examples

hmm.search(original.seq = fasta.file, regex.seq = PATT_REG, mafft.path = "complete path", hmm.path = "complete path")
}
\description{
This function uses MAFFT and HMMER to search for sequences with protien motifs using hidden markov models.
}
