% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.R
\name{hmm_build}
\alias{hmm_build}
\title{HMM build}
\usage{
hmm_build(hmm_path = NULL, seed = NULL, mafft_out = NULL)
}
\arguments{
\item{hmm_path}{complete path of HMMER}

\item{seed}{The seed to used with HMMER commands. Set this to get the same output each time}

\item{mafft_out}{MAFFT alignment output file name}
}
\value{
Resulted output file name

 @examples
hmm_build(hmm_path = NULL, seed = NULL,  mafft_out = NULL)
}
\description{
This function is build model by using MAFFT alignment sequence file
}
