% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.R
\name{HMM_fasta}
\alias{HMM_fasta}
\title{Generating HMM resulted fasta file}
\usage{
HMM_fasta(total_seq = NULL, original.seq = NULL)
}
\arguments{
\item{total_seq}{sequence containing (PATT_REG, HMM, HMM_table)}

\item{original.seq}{The absolute path for the original six-frame translation FASTA file.}
}
\value{
returns ouput fasta file name

 @examples
HMM_fasta(total_seq = NULL, original.seq = NULL)
}
\description{
This fuction is used for generating non redundant motif candidates fasta file.
}
