% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.R
\name{hmm_press}
\alias{hmm_press}
\title{Prepare a profile database for hmmscan}
\usage{
hmm_press(path_hmm = NULL)
}
\arguments{
\item{path_hmm}{complete path of HMMER}
}
\value{
Resulted output file name

 @examples
hmm_press <- function(path_hmm = NULL)
}
\description{
This function is Constructs binary compressed datafiles for hmmscan, starting from a profile database hmmfile in standard HMMER3 format.
}
