% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.R
\name{hmmbuild}
\alias{hmmbuild}
\title{HMM build}
\usage{
hmmbuild(hmm_path = NULL, seed = NULL, mafft.out.name = NULL)
}
\arguments{
\item{hmm_path}{complete path of HMMER}

\item{seed}{The seed to used with HMMER commands. Set this to get the same output each time.}

\item{mafft.out.name}{mafft alignment filename.}
}
\value{
Resulted output file name
}
\description{
This function is build model by using MAFFT alignment sequence file
}
\keyword{general}
