% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.R
\name{hmmpress}
\alias{hmmpress}
\title{Prepare a profile database for hmmscan}
\usage{
hmmpress(hmmpress_path = NULL)
}
\arguments{
\item{hmmpress_path}{complete path of HMMER}
}
\value{
Resulted output file name
}
\description{
This function is Constructs binary compressed datafiles for hmmscan, starting from a profile database hmmfile in standard HMMER3 format.
}
\keyword{general}
