% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.R
\name{hmmsearch}
\alias{hmmsearch}
\title{search profile(s) against a sequence database}
\usage{
hmmsearch(hmmsearch_path = NULL, seed = NULL, hmm.tresh = NULL,
  original.seq = NULL)
}
\arguments{
\item{hmmsearch_path}{complete path of HMMER}

\item{seed}{The seed to used with HMMER commands. Set this to get the same output each time.}

\item{hmm.tresh}{Set the bit score cutoff for the per-sequence ranked hit list to a real number. (Default = 0)}

\item{original.seq}{The absolute path for the original six-frame translation FASTA file}
}
\value{
Resulted output file name
}
\description{
hmmsearch is used to search one or more profiles against a sequence database.
}
\keyword{general}
