% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.R
\name{modify_filename}
\alias{modify_filename}
\title{modifying filenames and set default names}
\usage{
modify_filename(file = NULL, programe_name = NULL,
  concat_string = ".")
}
\arguments{
\item{file}{for input filename.}

\item{programe_name}{Programe name}

\item{concat_string}{used for concatination of string in to filename}
}
\value{
return output filename

 @examples

modify_filename(file = NULL, programe_name = NULL, concat_string = ".")
}
\description{
This function modifying filenames and set default names
.
}
