% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orfdiscard.R
\name{orf_discard}
\alias{orf_discard}
\title{Descard ORF file}
\usage{
orf_discard(orf_file = NULL, upper.limit = NULL, lower.limit = NULL,
  output_file = NULL)
}
\arguments{
\item{orf_file}{ORF file created from getorf function.}

\item{upper.limit}{enter sequnece length it will descard the sequences within specified range.}

\item{lower.limit}{enter sequence length.}

\item{output_file}{output filename defined by user.}
}
\value{
This function returns output file name

 @examples

orf_discard(orf_file= "filename", upper.limit= 1800, lower.limit= 100,  output_file = NULL)
}
\description{
This function descard the ORF protien sequences based on its length. The length cutoff is managed by lowerlimit and upperlimit.
}
