% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_file}
\alias{parse_file}
\title{Parses single fasta file into multiple files}
\usage{
parse_file(path_in = NULL, path_out = NULL, num_seq = NULL,
  trim = FALSE, trunc = 0, id = TRUE)
}
\arguments{
\item{path_in}{A comlpete path to the input fasta sequence file}

\item{path_out}{A Path where resulting files will be store.Path should be contain prefix name of input file on which will append .fa extenstion}

\item{num_seq}{Integer defining howmany sequences want to seperate in to one file}

\item{trim}{Logical, should the sequences be trimmed to 4000 amino acids to bypass the CBS server restrictions. Defaults to FALSE.}

\item{trunc}{Integer, truncate the sequences to this length.}

\item{id}{Logical, should the protein id's be returned. Defaults to FALSE.}
}
\value{
if id = FALSE, A Character vector of the paths to the resulting .FASTA formatted files.


 @examples
parse_file(path_in= "path of input file", path_out= "path to store outputfiles", num_seq= 300,trim = FALSE, trunc = 0, id = TRUE)
}
\description{
This function Parses the large fasta file and store the sequences into multiple files according to user requirement
}
