% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternsearch.R
\name{pattern.search}
\alias{pattern.search}
\title{Searching for motifs using regular expressions}
\usage{
pattern.search(fasta.file = NULL, reg.pat = NULL)
}
\arguments{
\item{fasta.file}{Provide protien sequence fasta file.}

\item{reg.pat}{Provide R regular expression pattern for searching motifs.}
}
\value{
A list contain motif sequences and motif dataframe

 @examples

 pattern.search(fasta.file = NULL, reg.pat = NULL)
}
\description{
This function searching the motif of interest by user defined regular expression against input fasta sequences.
}
