% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.R
\name{prepare_inputdata}
\alias{prepare_inputdata}
\title{input data preparation}
\usage{
prepare_inputdata(file_input = NULL, dir_name = NULL)
}
\arguments{
\item{file_input}{input file name}

\item{dir_name}{For storing parsing files}
}
\value{
file names


 @examples
prepare_inputdata(file_input = NULL, dir_name = NULL)
}
\description{
This function is used for preparation of input data
}
