% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.R
\name{prepare_inputdata_signalp}
\alias{prepare_inputdata_signalp}
\title{Preparation of input data to signalp}
\usage{
prepare_inputdata_signalp(version = NULL, file = NULL)
}
\arguments{
\item{version}{version of signalp programe.}

\item{file}{for input filename.}
}
\value{
return input filename


 @examples
prepare_inputdata_signalp(version = NULL, file = NULL)
}
\description{
This function is used for preparation of input data to signalp
.
}
