% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.R
\name{reading_hmm.results}
\alias{reading_hmm.results}
\title{Reading HMM resulted file}
\usage{
reading_hmm.results(original_seq = NULL, hmmsearch.out = NULL,
  hmmbuild.out = NULL, regex.seq = NULL, save.alignment = FALSE)
}
\arguments{
\item{original_seq}{The absolute path for the original six-frame translation FASTA file}

\item{hmmsearch.out}{hmmsearch output file}

\item{hmmbuild.out}{hmmbuild output file name}

\item{regex.seq}{pattern search function returns the sequence stored in PATT_REG variable.}

\item{save.alignment}{alignment file}
}
\value{
returns total sequence (HMM, PATT_REG)

 @examples
reading_hmm.results(original_seq = NULL, hmmsearch.out = NULL, hmmbuild.out = NULL, regex.seq = NULL, save.alignment = FALSE)
}
\description{
hmmsearch is used to search one or more profiles against a sequence database.
}
