% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.R
\name{run_mafft}
\alias{run_mafft}
\title{Running MAFFT Alignment}
\usage{
run_mafft(mafft_path = NULL, num_threads = NULL, file_name = NULL,
  mafft.out_file = NULL)
}
\arguments{
\item{mafft_path}{complete path of MAFFT}

\item{num_threads}{Number of threads to be used by MAFFT.}

\item{file_name}{PATT_REG sequence}

\item{mafft.out_file}{MAFFT ouputput filename}
}
\value{
Resulted output file name


 @examples
run_mafft(mafft_path = NULL, num_threads = NULL, file_name = NULL, mafft.out_file = NULL)
}
\description{
This function is perfoms multiple sequence alignment
}
