% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.R
\name{run_signalp}
\alias{run_signalp}
\title{Running signalp}
\usage{
run_signalp(programe_path = NULL, version = NULL,
  input_file.list = NULL, org.type = NULL, output = NULL)
}
\arguments{
\item{programe_path}{Full path of programe.}

\item{version}{version of signalp programe.}

\item{input_file.list}{input files}

\item{org.type}{For type of organism}

\item{output}{output file name}
}
\value{
return output filename


 @examples
run_signalp(programe_path =NULL, version = NULL, input_file.list = NULL, org.type = NULL, output = NULL)
}
\description{
This function is used for Running signalp
}
