% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.R
\name{run_targetp}
\alias{run_targetp}
\title{FOR running targetp}
\usage{
run_targetp(programe_path = NULL, input_file.list = NULL,
  org.group = NULL, output = NULL)
}
\arguments{
\item{programe_path}{Programe complete path}

\item{input_file.list}{Input filenames}

\item{org.group}{Organism grop name}

\item{output}{output filenam}
}
\value{
Resulted output file name


 @examples
run_targetp(programe_path =NULL, input_file.list = NULL, org.group = NULL, output = NULL)
}
\description{
This function is used forrun targetp
}
