% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.R
\name{run_tmhmm}
\alias{run_tmhmm}
\title{FOR running tmhmm}
\usage{
run_tmhmm(tmhmm.path = NULL, input_file.list = NULL, output = NULL)
}
\arguments{
\item{tmhmm.path}{Programe complete path}

\item{input_file.list}{Input filenames}

\item{output}{output filename}
}
\value{
Resulted output file name


 @examples
run_tmhmm(tmhmm.path = NULL, input_file.list = NULL, output = NULL)
}
\description{
This function is used forrun tmhmm
}
