% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.R
\name{siganlp5_summary}
\alias{siganlp5_summary}
\title{Dataframe for signalp}
\usage{
siganlp5_summary(signalp5out = NULL)
}
\arguments{
\item{signalp5out}{signalp version 5 output filename.}
}
\value{
return Datframe


 @examples
siganlp5_summary(signalp5out = NULL)
}
\description{
This function is used for producing Dataframe for signalp
}
