% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.R
\name{summary_signalp3}
\alias{summary_signalp3}
\title{Dataframe for signalp}
\usage{
summary_signalp3(signalp3_outname = NULL, original_input = NULL)
}
\arguments{
\item{signalp3_outname}{signalp version 3 output filename.}

\item{original_input}{orginal input file name.}
}
\value{
return Datframe


 @examples
summary_signalp3(signalp3_outname = NULL, original_input = NULL)
}
\description{
This function is used for producing Dataframe for signalp
}
