% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.R
\name{targetp_summary}
\alias{targetp_summary}
\title{Dataframe targetp}
\usage{
targetp_summary(targetp_outfile = NULL, organism = NULL,
  error = TRUE, verbose = FALSE)
}
\arguments{
\item{targetp_outfile}{Resulted targetp output filename}

\item{organism}{organism name (Plant or nonplant)}

\item{error}{If \code{TRUE}, throw an error if programs is not installed.}

\item{verbose}{If \code{TRUE}, print progress reports.}
}
\value{
Resulted output file name


 @examples
targetp_summary(targetp_outfile = NULL)
}
\description{
This function is used for producing Dataframe targetpp
}
