% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.R
\name{tmhmm_summary}
\alias{tmhmm_summary}
\title{Dataframe tmhmm}
\usage{
tmhmm_summary(tmhmm_outfile = NULL)
}
\arguments{
\item{tmhmm_outfile}{Resulted tmhmm output filename}
}
\value{
Resulted output file name


 @examples
tmhmm_summary(tmhmm_outfile = NULL)
}
\description{
This function is used for producing Dataframe tmhmm
}
